backend/requirements/base.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
razorpay==1.4.1
pydantic==2.6.1
django-storages==1.14.2
elasticsearch-dsl==7.4.1
sentry-sdk==1.39.1
django-prometheus==2.3.1
gunicorn==21.2.0
boto3==1.34.34
cryptography==42.0.2